cssom-view-1/src/CaretPosition-001.html

<!DOCTYPE html>
<html>
<head>
	<title>CSSOM View Module test:CaretPosition</title>
    <link rel="author" title="unbug" href="mailto:tidelgl@gmail.com" />
    <link rel="help" href="http://www.w3.org/TR/cssom-view/#the-caretposition-interface">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>	
	<style type="text/css">
	.box{
	    position: absolute;
	    top : 100px;
	    left : 400px;
	    width: 100px;
	    height: 100px;
	    background-color: #000;
	}
	</style>
</head>
<body>
	<p>This case tests the Screen pixelDepth and colorDepth</p>
	<p>The test passes if the value is Element "#box2"</p>
	<div id="box" class="box"></div>
	<div class="box"></div>
	<div class="box"></div>
	<div class="box"></div>
	<div class="box"></div>
	<div class="box"></div>
	<div id="box2" class="box"></div>
	<div id="box3" class="box" style="left: 405px;background:red;"></div>
	<div id="log"></div>
	<script>
        test(function(){
            assert_equals(getBox('box2'), document.elementFromPoint(400,100), "Expected value for element id is 'box2'");
        },'getBox');
        test(function(){
            assert_equals(null, document.elementFromPoint(400,900), "Expected value for element id is 'box2'");
        },'getBox');
        function getBox(id){
            return document.getElementById(id);
        }	    
	</script>
</body>
</html>
